moves network name validation to a new controller
